Summary
First, I made a constructor to assign the values to variables.  For the getFullName() method, I made it return the first name and last name. For the setFavoriteAnimal(), I assigned the input animal to the favoriteAnimal variable and returned that animal. For the happyBirthday() method, I printed out "Happy Birthday!" and incremented the age variable. I also printed out the age.